Add support for theme hbshelpers
This commit adds support for an hbshelpers
folder in the theme, which contains custom
hbs helpers used in the theme, with one
helper per file, with the helper name
being the name of the file.

J-SLAP-894
TEST=manual, auto

test that, after removing the babel and isNonRelativeUrl
helpers from jambo, and adding them to the themes hbshelpers
folder, that I can still build the page and it works in ie11
