+++ package.json
@@ -20,8 +20,6 @@
   },
   "homepage": "https://github.com/yext/jambo#readme",
   "dependencies": {
-    "@babel/core": "^7.9.6",
-    "@babel/preset-env": "^7.9.6",
     "comment-json": "^3.0.2",
     "file-system": "^2.2.2",
     "fs-extra": "^8.1.0",
@@ -38,7 +36,6 @@
     "yargs": "^15.1.0"
   },
   "devDependencies": {
-    "@babel/plugin-transform-object-assign": "^7.12.1",
     "eslint": "^7.8.1",
     "jest": "^25.4.0"
   },
