+++ src/errors/usererror.js
@@ -14,8 +14,6 @@ class UserError extends Error {
     } else {
       Error.captureStackTrace(this, this.constructor);
     }
-
-    
   }
 }
